+++ src/main/java/webserver/service/RequestHandler.java
@@ -1,0 +1,42 @@
+package webserver.service;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.io.OutputStream;
+import java.net.Socket;
+import java.nio.charset.StandardCharsets;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import webserver.domain.RequestLine;
+
+public class RequestHandler implements Runnable {
+    private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
+    private final Socket connection;
+
+    public RequestHandler(Socket connectionSocket) {
+        this.connection = connectionSocket;
+    }
+
+    public void run() {
+        logger.debug("New Client Connect! Connected IP : {}, Port : {}", connection.getInetAddress(),
+                connection.getPort());
+
+        try (InputStream in = connection.getInputStream(); OutputStream out = connection.getOutputStream()) {
+            BufferedReader br = new BufferedReader(new InputStreamReader(in, StandardCharsets.UTF_8));
+            String line = br.readLine();
+            RequestLine requestLine = RequestLine.create(line);
+            while (!"".equals(line)) {
+                System.out.println(line);
+                line = br.readLine();
+            }
+            ResponseHandler responseHandler = new ResponseHandler(connection);
+            responseHandler.run(requestLine);
+        } catch (IOException e) {
+            logger.error(e.getMessage());
+        }
+    }
+}
