+++ src/main/java/webserver/service/ResponseHandler.java
@@ -1,29 +1,25 @@
-package webserver;
+package webserver.service;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.Socket;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class RequestHandler implements Runnable {
+import webserver.domain.RequestLine;
+
+public class ResponseHandler {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
+    private final Socket connection;
 
-    private Socket connection;
-
-    public RequestHandler(Socket connectionSocket) {
-        this.connection = connectionSocket;
+    public ResponseHandler(Socket connection) {
+        this.connection = connection;
     }
 
-    public void run() {
-        logger.debug("New Client Connect! Connected IP : {}, Port : {}", connection.getInetAddress(),
-                connection.getPort());
-
-        try (InputStream in = connection.getInputStream(); OutputStream out = connection.getOutputStream()) {
-            // TODO 사용자 요청에 대한 처리는 이 곳에 구현하면 된다.
+    public void run(RequestLine requestLine) throws IOException {
+        try (OutputStream out = connection.getOutputStream()) {
             DataOutputStream dos = new DataOutputStream(out);
             byte[] body = "Hello World".getBytes();
             response200Header(dos, body.length);
